feat: Adds keyboard shortcut page and inline table edit
